gst-launch: fix padprobe duration error
This error happens randomly when playing video with padprobe, especially
in duration test.

The rootcause is using g_print inside a SIGALRM handler.
Every kind of "print" is non-reentrant so put it inside a signal
handler is not safe.

This patch fixes this issue by replacing g_print with write.

Signed-off-by: HungTran <hung.tran.jy@rvc.renesas.com>
Signed-off-by: Loc Vu <loc.vu.zn@renesas.com>

Porting from 1.16.3 to 1.22.11
Signed-off-by: duytruong <duy.truong.yw@renesas.com>